MultiLevelImage now extends SourcelessOpImage, which allows to remove it from cache. Remove AbstractNetcdfMultiLevelImage and NetcdfMultiLevelImage They were actually MultilevelSources and prevented to let MultiLevelImage extends SourcelessOpImage Implemented a more generic LazyMultiLevelSource which can be moved to ceres-glayer.
There will is a second PR for the opttbx to make use of the changes for the S3NetcdfMultiLevelImage
